backend/requirements.txt
fastapi[all]
uvicorn
websockets
pydantic
transformers
pika
python-dotenv
pytest
pytest-asyncio
httpx
redis
langid
langdetect
openai
# libretranslatepy